# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a ERP/POS API for G4H.",
        "title": "G4H ERP/POS API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "G4H ERP/POS API",
            "url": "https://github.com/aslon1213/g4h_pos_erp",
            "email": "hamidovaslon1@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/journals/{journal_id}/operations": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new transaction and update the journal",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals/operations"
                ],
                "summary": "Create a new operation transaction",
                "parameters": [
                    {
                        "description": "Transaction data",
                        "name": "transaction",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.JournalOperationInput"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Journal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
        "/api/v1/admin/journals/{journal_id}/operations/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Get an operation transaction by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals/operations"
                ],
                "summary": "Get an operation transaction by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Transaction"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an operation transaction by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals/operations"
                ],
                "summary": "Update an operation transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Amount",
                        "name": "amount",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Description",
                        "name": "description",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Journal"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete an operation transaction by ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "journals/operations"
                ],
                "summary": "Delete an operation transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Operation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Journal ID",
                        "name": "journal_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Output-models_Journal"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorOutput"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.Branch": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "phone": {
                    "type": "string"
                }
            }
        },
...
        "models.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "models.ErrorOutput": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {}
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
...
        "models.InitiatorType": {
            "type": "string",
            "enum": [
                "salary",
                "rent",
                "utilities",
                "other",
                "sale",
                "supplier",
                "bnpl"
            ],
            "x-enum-comments": {
                "InitiatorTypeBNPL": "buy now pay later BNPL transactions"
            },
            "x-enum-varnames": [
                "InitiatorTypeSalary",
                "InitiatorTypeRent",
                "InitiatorTypeUtilities",
                "InitiatorTypeOther",
                "InitiatorTypeSales",
                "InitiatorTypeSupplier",
                "InitiatorTypeBNPL"
            ]
        },
        "models.Journal": {
            "type": "object",
            "properties": {
                "branch": {
                    "$ref": "#/definitions/models.Branch"
                },
                "cash_left": {
                    "type": "integer"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "operations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "shift_is_closed": {
                    "type": "boolean"
                },
                "terminal_income": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "models.JournalOperationInput": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "supplier_id": {
                    "type": "string"
                },
                "supplier_transaction": {
                    "type": "boolean"
                },
                "type": {
                    "$ref": "#/definitions/models.TransactionType"
                }
            }
        },
...
        "models.Output-models_Journal": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Journal"
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
...
        "models.Output-models_Transaction": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/models.Transaction"
                },
                "error": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Error"
                    }
                }
            }
        },
        "models.PaymentMethod": {
            "type": "string",
            "enum": [
                "cash",
                "bank",
                "terminal",
                "online_payment",
                "cheque",
                "online_transfer",
                "undefined"
            ],
            "x-enum-varnames": [
                "PaymentMethodCash",
                "PaymentMethodBank",
                "PaymentMethodTerminal",
                "OnlineMobileAppPayment",
                "Cheque",
                "OnlineTransfer",
                "PaymentMethodUndefined"
            ]
        },
...
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "branch_id": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "payment_method": {
                    "$ref": "#/definitions/models.PaymentMethod"
                },
                "type": {
                    "$ref": "#/definitions/models.InitiatorType"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "models.TransactionType": {
            "type": "string",
            "enum": [
                "credit",
                "debit"
            ],
            "x-enum-comments": {
                "TransactionTypeCredit": "credit means - income - when money is gained or received into an account",
                "TransactionTypeDebit": "debit means - outcome - when money is lost, spent, or withdrawn from an account"
            },
            "x-enum-varnames": [
                "TransactionTypeCredit",
                "TransactionTypeDebit"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}